+++ rumbas/src/data/extension.rs
@@ -23,3 +23,29 @@ impl ToNumbas for Extensions {
         }
     }
 }
+
+impl Extensions {
+    pub fn new() -> Extensions {
+        Extensions {
+            jsx_graph: Value::Normal(false),
+        }
+    }
+
+    pub fn combine(e: Extensions, f: Extensions) -> Extensions {
+        Extensions {
+            jsx_graph: Value::Normal(e.jsx_graph.unwrap() || f.jsx_graph.unwrap()),
+        }
+    }
+
+    pub fn to_paths(&self) -> Vec<String> {
+        let numbas_path = std::env::var("NUMBAS_FOLDER").expect("NUMBAS_FOLDER to be set"); //TODO: static str for NUMBAS_FOLDER
+        let mut paths = Vec::new();
+        if self.jsx_graph.unwrap() {
+            paths.push("extensions/jsxgraph");
+        }
+        paths
+            .into_iter()
+            .map(|s| format!("{}/{}", numbas_path, s))
+            .collect()
+    }
+}